Move types outside components folder
