+++ src/components/Landing/components/DropdownMenu.tsx
@@ -1,7 +1,7 @@
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faBars } from '@fortawesome/free-solid-svg-icons';
 import React, { useEffect, useRef, useState } from 'react';
-import { Button } from './types';
+import { Button } from '../types';
 
 interface Props {
     buttonsList: Button[];
