+++ src/components/Landing/types.ts
